Revert "Ticket:42  Add Create Category Feature"
