+++ src/views/ApplicationViews.js
@@ -3,14 +3,13 @@ import { Login } from "../components/auth/Login"
 import { Register } from "../components/auth/Register"
 import { Authorized } from "./Authorized"
 import { TagManager } from "../components/Tags/TagManager.js"
+import { CategoryList } from "../categories/CategoryList.js"
 import { AllPosts } from "../components/posts/AllPosts"
 import React, { useEffect, useState } from "react"
 import { getAllPosts, getUserPosts } from "../managers/PostManager"
 import { UserPosts } from "../components/posts/UserPosts.js"
 import { CreatePost } from "../components/posts/CreatePost.js"
-import { CategoryList } from "../components/categories/CategoryList.js"
 import { PostDetails } from "../components/posts/PostDetails.js"
-
 
 export const ApplicationViews = ({ token, setToken }) => {
